group招待機能を修正

+++ app/controllers/user_group_relations_controller.rb
@@ -6,14 +6,18 @@ class UserGroupRelationsController < ApplicationController
   end
 
   def create
-    group = Group.find(create_params[:group_id])
     user = User.find(create_params[:user_id])
-    request = Request.find_by(create_params)
-    if current_user.in_this_group?(group)
-      group.users << user
-      request.destroy unless request.nil?
+    unless create_params[:group_id].blank?
+      group = Group.find(create_params[:group_id])
+      request = Request.find_by(create_params)
+      if current_user.in_this_group?(group)
+        group.users << user
+        request.destroy unless request.nil?
+      end
+      redirect_to group_path(group)
+    else
+      redirect_to user_path(user)
     end
-    redirect_to group_path(group)
   end
 
   private
